# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for AutoBan.",
        "title": "AutoBan API",
        "contact": {},
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/vehicles/{vehicle_id}/service-visits": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all service visit records for a specific user vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "List service visits for a user vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListServiceVisitsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new service visit record for a user vehicle with optional services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "Create a new service visit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service visit data",
                        "name": "service_visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateServiceVisitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceVisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/service-visits/last": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the most recent service visit record for a specific user vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "Get last service visit for a user vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceVisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/service-visits/{visit_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific service visit record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "Get service visit by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service visit ID",
                        "name": "visit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceVisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing service visit record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "Update service visit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service visit ID",
                        "name": "visit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated service visit data",
                        "name": "service_visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateServiceVisitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceVisitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a service visit record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Visits"
                ],
                "summary": "Delete service visit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Service visit ID",
                        "name": "visit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateServiceVisitRequest": {
            "description": "Service visit creation request",
            "type": "object",
            "required": [
                "service_date",
                "service_mileage",
                "user_vehicle_id"
            ],
            "properties": {
                "notes": {
                    "description": "Additional notes",
                    "type": "string",
                    "example": "Regular maintenance service"
                },
                "oil_change": {
                    "description": "Oil change information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ServiceVisitOilChange"
                        }
                    ]
                },
                "oil_filter": {
                    "description": "Oil filter information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ServiceVisitOilFilter"
                        }
                    ]
                },
                "service_center": {
                    "description": "Service center where service was performed",
                    "type": "string",
                    "example": "Auto Service Center"
                },
                "service_date": {
                    "description": "Date when service was performed",
                    "type": "string",
                    "example": "2024-01-15"
                },
                "service_mileage": {
                    "description": "Mileage when service was performed",
                    "type": "integer",
                    "minimum": 0,
                    "example": 15000
                },
                "user_vehicle_id": {
                    "description": "ID of the user vehicle",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "dto.ListServiceVisitsResponse": {
            "description": "List of service visits",
            "type": "object",
            "properties": {
                "service_visits": {
                    "description": "List of service visits",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceVisitResponse"
                    }
                }
            }
        },
...
        "dto.ServiceVisitOilChange": {
            "description": "add oil change to create service visit",
            "type": "object",
            "required": [
                "oil_name"
            ],
            "properties": {
                "next_change_date": {
                    "description": "Next change date",
                    "type": "string",
                    "example": "2021-01-01"
                },
                "next_change_mileage": {
                    "description": "Next change mileage",
                    "type": "integer",
                    "minimum": 0,
                    "example": 20000
                },
                "notes": {
                    "description": "Notes",
                    "type": "string",
                    "example": "تعویض روغن"
                },
                "oil_brand": {
                    "description": "Oil brand",
                    "type": "string",
                    "example": "بهران"
                },
                "oil_capacity": {
                    "description": "Oil capacity",
                    "type": "number",
                    "example": 5
                },
                "oil_name": {
                    "description": "Oil name",
                    "type": "string",
                    "example": "تکتاز"
                },
                "oil_type": {
                    "description": "Oil type",
                    "type": "string",
                    "example": "مینرال، سنتتیک، نیمه سنتتیک"
                },
                "oil_viscosity": {
                    "description": "Oil viscosity",
                    "type": "string",
                    "example": "5W-30, 10W-40, etc."
                }
            }
        },
...
        "dto.ServiceVisitOilFilter": {
            "description": "Requst to add oil filter to create service visit",
            "type": "object",
            "required": [
                "filter_name"
            ],
            "properties": {
                "filter_brand": {
                    "description": "Brand of the oil filter",
                    "type": "string",
                    "example": "Mann"
                },
                "filter_name": {
                    "description": "Name of the oil filter",
                    "type": "string",
                    "example": "Oil Filter"
                },
                "filter_part_number": {
                    "description": "Part number of the oil filter",
                    "type": "string",
                    "example": "HU816x"
                },
                "filter_type": {
                    "description": "Type of the oil filter",
                    "type": "string",
                    "example": "Cartridge"
                },
                "next_change_date": {
                    "description": "Next change date",
                    "type": "string",
                    "example": "2024-07-15"
                },
                "next_change_mileage": {
                    "description": "Next change mileage",
                    "type": "integer",
                    "minimum": 0,
                    "example": 30000
                },
                "notes": {
                    "description": "Additional notes",
                    "type": "string",
                    "example": "Changed with oil change"
                }
            }
        },
...
        "dto.ServiceVisitResponse": {
            "description": "Service visit response",
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the service visit",
                    "type": "string"
                },
                "notes": {
                    "description": "Additional notes",
                    "type": "string"
                },
                "oil_change": {
                    "description": "Oil change information (if performed)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ServiceVisitOilChangeResponse"
                        }
                    ]
                },
                "oil_filter": {
                    "description": "Oil filter information (if performed)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ServiceVisitOilFilterResponse"
                        }
                    ]
                },
                "service_center": {
                    "description": "Service center where service was performed",
                    "type": "string"
                },
                "service_date": {
                    "description": "Date when service was performed",
                    "type": "string"
                },
                "service_mileage": {
                    "description": "Mileage when service was performed",
                    "type": "integer"
                },
                "user_vehicle_id": {
                    "description": "ID of the user vehicle",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateServiceVisitRequest": {
            "description": "Service visit update request",
            "type": "object",
            "properties": {
                "notes": {
                    "description": "Additional notes",
                    "type": "string",
                    "example": "Regular maintenance service"
                },
                "oil_change": {
                    "description": "Oil change information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UpdateServiceVisitOilChange"
                        }
                    ]
                },
                "oil_filter": {
                    "description": "Oil filter information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.UpdateServiceVisitOilFilter"
                        }
                    ]
                },
                "service_center": {
                    "description": "Service center where service was performed",
                    "type": "string",
                    "example": "Auto Service Center"
                },
                "service_date": {
                    "description": "Date when service was performed",
                    "type": "string",
                    "example": "2024-01-15"
                },
                "service_mileage": {
                    "description": "Mileage when service was performed",
                    "type": "integer",
                    "minimum": 0,
                    "example": 15000
                }
            }
        },
...
        "errors.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable code (stable across languages)",
                    "type": "string"
                },
                "details": {
                    "description": "Optional structured details for debugging (safe to expose)",
                    "type": "object",
                    "additionalProperties": {}
                },
                "fields": {
                    "description": "Optional list of field-level errors (validation)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errors.FieldError"
                    }
                },
                "message": {
                    "description": "Bilingual message for human-readable display",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorMessage"
                        }
                    ]
                },
                "retryable": {
                    "description": "Whether the operation can be retried by client later",
                    "type": "boolean"
                }
            }
        },
        "errors.ErrorMessage": {
            "type": "object",
            "properties": {
                "english": {
                    "type": "string"
                },
                "persian": {
                    "type": "string"
                }
            }
        },
        "errors.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/errors.ErrorMessage"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}